Refactoring filter settings to work with singleton pattern
