+++ Cheers/Model/Singletons/ColorsSingleton.swift
