+++ Cheers/Model/Singletons/FavoritesSingleton.swift
