+++ Cheers/Model/Singletons/SettingsSingleton.swift
